backend/requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Mail==0.10.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
qrcode==8.2
redis==6.2.0
Werkzeug==3.1.3
SQLAlchemy==2.0.32
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
pytest==8.3.4
pytest-flask==1.3.0
flask-socketio==5.3.6
